service-portal file restructure
