+++ libs/service-portal/signature-collection/src/hooks/graphql/mutations.ts
