+++ libs/service-portal/signature-collection/src/screens/CandidateView/CancelCollection/index.tsx
@@ -1,12 +1,12 @@
 import { Box, Button, Text, toast } from '@island.is/island-ui/core'
 import { useLocale, useNamespaces } from '@island.is/localization'
-import { m } from '../../lib/messages'
+import { m } from '../../../lib/messages'
 import { Modal } from '@island.is/service-portal/core'
 import { useState } from 'react'
-import { useGetListsForUser, useIsOwner } from '../hooks'
+import { useGetListsForUser, useIsOwner } from '../../../hooks'
 import { useMutation } from '@apollo/client'
-import { SignatureCollectionSuccess } from '../../types/schema'
-import { cancelCollectionMutation } from '../mutations'
+import { cancelCollectionMutation } from '../../../hooks/graphql/mutations'
+import { SignatureCollectionSuccess } from '@island.is/api/schema'
 
 const CancelCollection = ({ collectionId }: { collectionId: string }) => {
   useNamespaces('sp.signatureCollection')
