+++ libs/service-portal/signature-collection/src/screens/CandidateView/ViewList/index.tsx
@@ -1,9 +1,9 @@
 import { Box, Stack, Text } from '@island.is/island-ui/core'
 import { useLocale, useNamespaces } from '@island.is/localization'
-import { m } from '../../lib/messages'
+import { m } from '../../../lib/messages'
 import Signees from './signees'
 import { useLocation } from 'react-router-dom'
-import { useGetSignatureList } from '../hooks'
+import { useGetSignatureList } from '../../../hooks'
 import format from 'date-fns/format'
 
 const ViewList = () => {
