+++ libs/service-portal/signature-collection/src/screens/SigneeView/index.tsx
@@ -10,11 +10,11 @@ import { useLocale, useNamespaces } from '@island.is/localization'
 import { m } from '../../lib/messages'
 import { IntroHeader, Modal } from '@island.is/service-portal/core'
 import { useState } from 'react'
-import { useGetListsForUser, useGetSignedList } from '../hooks'
+import { useGetListsForUser, useGetSignedList } from '../../hooks'
 import format from 'date-fns/format'
-import { Skeleton } from '../Skeletons'
+import { Skeleton } from '../skeletons'
 import { useMutation } from '@apollo/client'
-import { unSignList } from '../mutations'
+import { unSignList } from '../../hooks/graphql/mutations'
 import { SignatureCollectionSuccess } from '@island.is/api/schema'
 
 const SigneeView = () => {
