+++ libs/service-portal/signature-collection/src/screens/skeletons.tsx
